Add geo location service

+++ src/main/java/com/propertyviewerproject/propertyviewer/entity/Building.java
@@ -28,7 +28,7 @@ public class Building {
     @JoinColumn(name = "USER_ID")
     private User user;
 
-    private int lon;
-    private int lat;
+    private double lon;
+    private double lat;
 
 }
